fix: preserve multipart upload snapshots and cached subtypes
