+++ src/internal/uploads.ts
@@ -2,6 +2,8 @@ import type { RequestOptions } from './request-options';
 import type { FilePropertyBag, Fetch } from './builtin-types';
 import type { OpenAI } from '../client';
 import { buildHeaders } from './headers';
+import type { MultipartDataSnapshot } from './multipart-snapshot';
+import { snapshotBlobData, snapshotResponseData, snapshotStreamingFileData } from './multipart-snapshot';
 import { makeReadableStream, ReadableStreamFrom, ReadableStreamToAsyncIterable } from './shims';
 import type { ReadableStream } from './shim-types';
 import { encodeUTF8 } from './utils/bytes';
@@ -158,18 +160,32 @@ export function makeFile(
  * discard their directories.
  */
 export function getName(value: any, options?: { stripFilename?: boolean | undefined }): string | undefined {
+  return getNameInternal(value, options);
+}
+
+function getNameInternal(
+  value: unknown,
+  options?: { stripFilename?: boolean | undefined },
+  knownNamedBlob = false,
+): string | undefined {
   if (typeof value !== 'object' || value === null) {
     return undefined;
   }
 
+  const candidate = value as {
+    name?: unknown;
+    filename?: unknown;
+    url?: unknown;
+    path?: unknown;
+  };
   const explicitName =
-    ('name' in value && value.name && String(value.name)) ||
-    ('filename' in value && value.filename && String(value.filename));
+    ((knownNamedBlob || 'name' in candidate) && candidate.name ? String(candidate.name) : undefined) ||
+    ('filename' in candidate && candidate.filename ? String(candidate.filename) : undefined);
   if (explicitName) {
     return options?.stripFilename === false ? normalizeFilenamePath(explicitName) : basename(explicitName);
   }
 
-  const url = 'url' in value && value.url && String(value.url);
+  const url = 'url' in candidate && candidate.url ? String(candidate.url) : undefined;
   if (url) {
     try {
       return basename(new URL(url).pathname);
@@ -178,7 +194,7 @@ export function getName(value: any, options?: { stripFilename?: boolean | undefi
     }
   }
 
-  const path = 'path' in value && value.path && String(value.path);
+  const path = 'path' in candidate && candidate.path ? String(candidate.path) : undefined;
   return path ? basename(path) : undefined;
 }
 
@@ -343,7 +359,7 @@ const isReadableStream = (value: unknown): value is ReadableStream<BlobPart> =>
 const isStreamingFile = (value: unknown): value is StreamingFile =>
   typeof value === 'object' && value !== null && brand_privateStreamingFile in value;
 
-type UploadableKind = 'upload' | 'streaming-upload' | 'streaming-file' | undefined;
+type UploadableKind = 'named-blob' | 'response' | 'streaming-upload' | 'streaming-file' | undefined;
 type UploadableKinds = WeakMap<object, UploadableKind>;
 
 const getUploadableKind = (value: unknown, uploadableKinds: UploadableKinds): UploadableKind => {
@@ -376,8 +392,10 @@ const getUploadableKind = (value: unknown, uploadableKinds: UploadableKinds): Up
     uploadableKind = 'streaming-file';
   } else if (isAsyncIterable(value) || isReadableStream(value)) {
     uploadableKind = 'streaming-upload';
-  } else if (value instanceof Response || isNamedBlob(value)) {
-    uploadableKind = 'upload';
+  } else if (value instanceof Response) {
+    uploadableKind = 'response';
+  } else if (isNamedBlob(value)) {
+    uploadableKind = 'named-blob';
   }
 
   if (uploadableKind) {
@@ -396,7 +414,7 @@ const hasStreamingUploadableValue = (value: unknown, uploadableKinds: Uploadable
   if (uploadableKind === 'streaming-file' || uploadableKind === 'streaming-upload') {
     return true;
   }
-  if (uploadableKind === 'upload') {
+  if (uploadableKind === 'named-blob' || uploadableKind === 'response') {
     return false;
   }
 
@@ -447,246 +465,6 @@ type FormEntry =
     }>;
 
 type MultipartEntry = FormEntry;
-
-type MultipartDataSnapshot = Readonly<{
-  data: StreamingFileInput;
-  dispose?: (() => void) | undefined;
-}>;
-
-function isNativeReadableStream(value: StreamingFileInput, requireLocked = false): boolean {
-  if (typeof globalThis.ReadableStream !== 'function') {
-    return false;
-  }
-
-  try {
-    const getLocked = Object.getOwnPropertyDescriptor(globalThis.ReadableStream.prototype, 'locked')?.get;
-    const locked = getLocked?.call(value);
-    return typeof locked === 'boolean' && (!requireLocked || locked);
-  } catch {
-    // Ordinary async iterables and reader-like objects do not satisfy the native stream brand.
-    return false;
-  }
-}
-
-async function ignoreCleanupResult(cleanup: () => unknown): Promise<void> {
-  try {
-    await cleanup();
-  } catch {
-    // Cleanup failures must not mask the primary multipart result.
-  }
-}
-
-function snapshotIteratorReturn(iterator: AsyncIterator<BlobPart>): AsyncIterator<BlobPart>['return'] {
-  try {
-    let prototype: object | null = iterator;
-    while (prototype !== null) {
-      const descriptor = Object.getOwnPropertyDescriptor(prototype, 'return');
-      if (descriptor) {
-        if ('value' in descriptor) {
-          const returnIterator = descriptor.value as AsyncIterator<BlobPart>['return'];
-          return returnIterator
-            ? (...args: [] | [unknown]) => Reflect.apply(returnIterator, iterator, args)
-            : undefined;
-        }
-
-        const getReturn = descriptor.get;
-        if (!getReturn) {
-          return undefined;
-        }
-
-        return (...args: [] | [unknown]) => {
-          const returnIterator = Reflect.apply(getReturn, iterator, []) as AsyncIterator<BlobPart>['return'];
-          return returnIterator
-            ? Reflect.apply(returnIterator, iterator, args)
-            : Promise.resolve({ done: true as const, value: args[0] });
-        };
-      }
-      prototype = Object.getPrototypeOf(prototype);
-    }
-  } catch (error) {
-    return () => {
-      throw error;
-    };
-  }
-
-  return (...args: [] | [unknown]) => {
-    const returnIterator = Reflect.get(iterator, 'return') as AsyncIterator<BlobPart>['return'];
-    return returnIterator
-      ? Reflect.apply(returnIterator, iterator, args)
-      : Promise.resolve({ done: true as const, value: args[0] });
-  };
-}
-
-function snapshotStreamingFileData(
-  value: StreamingFileInput,
-  snapshots: WeakMap<object, MultipartDataSnapshot>,
-): MultipartDataSnapshot {
-  const cached = snapshots.get(value);
-  if (cached) {
-    return cached;
-  }
-
-  const { [Symbol.asyncIterator]: createIterator } = value as AsyncIterable<BlobPart>;
-  if (typeof createIterator === 'function') {
-    const iterator = createIterator.call(value);
-    const returnIterator = snapshotIteratorReturn(iterator);
-    const isLockedNativeStream = isNativeReadableStream(value, true);
-    let next: AsyncIterator<BlobPart>['next'];
-    try {
-      ({ next } = iterator);
-    } catch (error) {
-      void ignoreCleanupResult(() => returnIterator?.());
-      throw error;
-    }
-    let consumed = false;
-    const snapshot: MultipartDataSnapshot = {
-      data: {
-        [Symbol.asyncIterator]() {
-          consumed = true;
-          return {
-            next(...args: [] | [undefined]) {
-              return next.call(iterator, ...args);
-            },
-            return(...args: [] | [unknown]) {
-              return returnIterator
-                ? returnIterator(...args)
-                : Promise.resolve({ done: true as const, value: args[0] });
-            },
-            [Symbol.asyncIterator]() {
-              return this;
-            },
-          };
-        },
-      },
-      dispose() {
-        if (!consumed) {
-          consumed = true;
-          void ignoreCleanupResult(() => returnIterator?.());
-        }
-      },
-    };
-    if (isLockedNativeStream) {
-      snapshots.set(value, snapshot);
-    }
-    return snapshot;
-  }
-
-  const { getReader } = value as ReadableStream<BlobPart>;
-  if (typeof getReader === 'function') {
-    const reader = getReader.call(value) as ReadableStreamDefaultReader<BlobPart>;
-    let read: ReadableStreamDefaultReader<BlobPart>['read'];
-    try {
-      ({ read } = reader);
-    } catch (error) {
-      void ignoreCleanupResult(() => reader.cancel());
-      try {
-        reader.releaseLock();
-      } catch {
-        // Cleanup failures must not mask the primary multipart result.
-      }
-      throw error;
-    }
-
-    let cancel: ReadableStreamDefaultReader<BlobPart>['cancel'];
-    try {
-      cancel = reader.cancel.bind(reader);
-    } catch (error) {
-      cancel = () => {
-        throw error;
-      };
-    }
-
-    let releaseLock: ReadableStreamDefaultReader<BlobPart>['releaseLock'];
-    try {
-      releaseLock = reader.releaseLock.bind(reader);
-    } catch (error) {
-      releaseLock = () => {
-        throw error;
-      };
-    }
-
-    const capturedReader = {
-      read: () => read.call(reader),
-      cancel,
-      releaseLock,
-    };
-    let consumed = false;
-    const snapshot: MultipartDataSnapshot = {
-      data: {
-        async *[Symbol.asyncIterator]() {
-          if (consumed) {
-            return;
-          }
-          consumed = true;
-          yield* ReadableStreamToAsyncIterable<BlobPart>({ getReader: () => capturedReader });
-        },
-      },
-      dispose() {
-        if (!consumed) {
-          consumed = true;
-          void ignoreCleanupResult(cancel);
-          try {
-            releaseLock();
-          } catch {
-            // Cleanup failures must not mask the primary multipart result.
-          }
-        }
-      },
-    };
-    if (isNativeReadableStream(value, true)) {
-      snapshots.set(value, snapshot);
-    }
-    return snapshot;
-  }
-
-  throw new TypeError('Streaming file data must be an async iterable or readable stream');
-}
-
-function snapshotBlobData(
-  value: Blob,
-  snapshots: WeakMap<object, MultipartDataSnapshot>,
-): MultipartDataSnapshot {
-  const { stream } = value as Blob & { stream?: Blob['stream'] };
-  if (typeof stream === 'function') {
-    return snapshotStreamingFileData(stream.call(value) as ReadableStream<BlobPart>, snapshots);
-  }
-
-  const immutableBlob = Blob.prototype.slice.call(value);
-  const { arrayBuffer: readImmutableBlob } = Blob.prototype;
-  const { arrayBuffer: read } = value;
-  return {
-    data: {
-      async *[Symbol.asyncIterator]() {
-        if (read === readImmutableBlob) {
-          yield await read.call(value);
-        } else {
-          await read.call(value);
-          yield await readImmutableBlob.call(immutableBlob);
-        }
-      },
-    },
-  };
-}
-
-function snapshotResponseData(
-  value: Response,
-  snapshots: WeakMap<object, MultipartDataSnapshot>,
-): MultipartDataSnapshot {
-  const { body } = value;
-  if (body) {
-    return snapshotStreamingFileData(body, snapshots);
-  }
-
-  const blob = value.blob();
-  void ignoreCleanupResult(() => blob);
-  return {
-    data: {
-      async *[Symbol.asyncIterator]() {
-        yield await blob;
-      },
-    },
-  };
-}
 
 const createStreamingFormRequestOptions = (
   opts: RequestOptions,
@@ -959,18 +737,21 @@ const addFormValue = async (
     );
   }
 
+  const cachedUploadableKind = typeof value === 'object' ? uploadableKinds?.get(value) : undefined;
+
   // Nested form keys use the current bracketed encoding.
   if (typeof value === 'string' || typeof value === 'number' || typeof value === 'boolean') {
     form.append(key, String(value));
-  } else if (value instanceof Response) {
+  } else if (cachedUploadableKind === 'response' || value instanceof Response) {
+    const response = value as Response;
     form.append(
       key,
-      makeFile([await value.blob()], getName(value, { stripFilename: options.stripFilenames })),
+      makeFile([await response.blob()], getName(response, { stripFilename: options.stripFilenames })),
     );
-  } else if (
-    !(uploadableKinds?.get(value as object) === 'upload' && value instanceof Blob) &&
-    isAsyncIterable(value)
-  ) {
+  } else if (cachedUploadableKind === 'named-blob') {
+    const blob = value as NamedBlob;
+    form.append(key, blob, getNameInternal(blob, { stripFilename: options.stripFilenames }, true));
+  } else if (isAsyncIterable(value)) {
     form.append(
       key,
       makeFile(
